Introduce QuizCache class to encapsulate the quiz cache behavior
