+++ src/main/java/de/tum/in/www1/artemis/service/scheduled/quiz/QuizScheduleService.java
@@ -7,7 +7,6 @@ import java.time.ZonedDateTime;
 import java.util.*;
 import java.util.Map.Entry;
 import java.util.concurrent.TimeUnit;
-import java.util.function.UnaryOperator;
 
 import javax.validation.constraints.NotNull;
 
@@ -19,12 +18,10 @@ import org.springframework.context.event.EventListener;
 import org.springframework.messaging.simp.SimpMessageSendingOperations;
 import org.springframework.stereotype.Service;
 
-import com.hazelcast.config.*;
+import com.hazelcast.config.Config;
 import com.hazelcast.core.HazelcastInstance;
 import com.hazelcast.cp.IAtomicReference;
-import com.hazelcast.map.IMap;
 import com.hazelcast.scheduledexecutor.*;
-import com.hazelcast.topic.ITopic;
 
 import de.tum.in.www1.artemis.config.Constants;
 import de.tum.in.www1.artemis.domain.Result;
@@ -50,12 +47,6 @@ public class QuizScheduleService {
 
     private static final String HAZELCAST_PROCESS_CACHE_HANDLER = QuizProcessCacheTask.HAZELCAST_PROCESS_CACHE_TASK + "-handler";
 
-    private static final String HAZELCAST_CACHED_EXERCISE_UPDATE_TOPIC = Constants.HAZELCAST_QUIZ_PREFIX + "cached-exercise-invalidation";
-
-    private IMap<Long, QuizExerciseCache> cachedQuizExercises;
-
-    private ITopic<QuizExercise> cachedQuizExerciseUpdates;
-
     private volatile IScheduledExecutorService threadPoolTaskScheduler;
 
     private IAtomicReference<ScheduledTaskHandler> scheduledProcessQuizSubmissions;
@@ -74,7 +65,7 @@ public class QuizScheduleService {
 
     private SimpMessageSendingOperations messagingTemplate;
 
-    private HazelcastInstance hazelcastInstance;
+    private QuizCache quizCache;
 
     public QuizScheduleService(SimpMessageSendingOperations messagingTemplate, StudentParticipationRepository studentParticipationRepository, ResultRepository resultRepository,
             UserService userService, QuizSubmissionRepository quizSubmissionRepository, HazelcastInstance hazelcastInstance) {
@@ -83,12 +74,8 @@ public class QuizScheduleService {
         this.resultRepository = resultRepository;
         this.userService = userService;
         this.quizSubmissionRepository = quizSubmissionRepository;
-        this.hazelcastInstance = hazelcastInstance;
         this.scheduledProcessQuizSubmissions = hazelcastInstance.getCPSubsystem().getAtomicReference(HAZELCAST_PROCESS_CACHE_HANDLER);
-        this.cachedQuizExercises = hazelcastInstance.getMap(Constants.HAZELCAST_EXERCISE_CACHE);
         this.threadPoolTaskScheduler = hazelcastInstance.getScheduledExecutorService(Constants.HAZELCAST_QUIZ_SCHEDULER);
-        this.cachedQuizExerciseUpdates = hazelcastInstance.getTopic(HAZELCAST_CACHED_EXERCISE_UPDATE_TOPIC);
-        this.cachedQuizExerciseUpdates.addMessageListener(newQuizExerciseMessage -> updateQuizExerciseLocally(newQuizExerciseMessage.getMessageObject()));
     }
 
     /**
@@ -97,22 +84,9 @@ public class QuizScheduleService {
      * @param config the {@link Config} the QuizScheduleService-specific configuration should be added to
      */
     public static void configureHazelcast(Config config) {
-        QuizExerciseCache.registerSerializers(config);
+        QuizCache.configureHazelcast(config);
         // Pool size default 16, increased capacity (as we could have many quizzes) and default durability for now
         config.getScheduledExecutorConfig(Constants.HAZELCAST_QUIZ_SCHEDULER).setPoolSize(16).setCapacity(1000).setDurability(1);
-        // Important to avoid continuous serialization and de-serialization and the implications on transient fields of QuizExerciseCache
-        EvictionConfig evictionConfig = new EvictionConfig() //
-                .setEvictionPolicy(EvictionPolicy.NONE);
-        NearCacheConfig nearCacheConfig = new NearCacheConfig() //
-                .setName(Constants.HAZELCAST_EXERCISE_CACHE + "-local") //
-                .setInMemoryFormat(InMemoryFormat.OBJECT) //
-                .setSerializeKeys(true) //
-                .setInvalidateOnChange(true) //
-                .setTimeToLiveSeconds(0) //
-                .setMaxIdleSeconds(0) //
-                .setEvictionConfig(evictionConfig) //
-                .setCacheLocalEntries(true);
-        config.getMapConfig(Constants.HAZELCAST_EXERCISE_CACHE).setNearCacheConfig(nearCacheConfig);
     }
 
     @EventListener(ApplicationReadyEvent.class)
@@ -134,84 +108,6 @@ public class QuizScheduleService {
     }
 
     /**
-     * Only for reading from QuizExerciseCache
-     *
-     * @param quizExerciseId the id of the quiz exercise, must not be null
-     */
-    private QuizExerciseCache getReadCacheFor(Long quizExerciseId) {
-        return cachedQuizExercises.getOrDefault(quizExerciseId, QuizExerciseCache.empty());
-    }
-
-    /**
-     * Only for the modification of transient properties, e.g. the exercise and the maps.
-     * <p>
-     * Creates new QuizExerciseCache if required.
-     *
-     * @param quizExerciseId the id of the quiz exercise, must not be null
-     */
-    private QuizExerciseCache getTransientWriteCacheFor(Long quizExerciseId) {
-        var cachedQuiz = cachedQuizExercises.get(quizExerciseId);
-        if (cachedQuiz != null)
-            return cachedQuiz;
-        cachedQuizExercises.lock(quizExerciseId);
-        try {
-            cachedQuiz = cachedQuizExercises.get(quizExerciseId);
-            if (cachedQuiz != null)
-                return cachedQuiz;
-            var newCachedQuiz = new QuizExerciseDistributedCache(quizExerciseId);
-            newCachedQuiz.setHazelcastInstance(hazelcastInstance);
-            cachedQuizExercises.set(quizExerciseId, newCachedQuiz);
-            // return the new deserialized, cached object returned by get()
-            // (this is not the newCachedQuiz anymore although we use near caching in OBJECT in-memory format, because Hazelcast.)
-            return cachedQuizExercises.get(quizExerciseId);
-        }
-        finally {
-            cachedQuizExercises.unlock(quizExerciseId);
-        }
-    }
-
-    /**
-     * To perform the given action on QuizExerciseCache non-transient fields.
-     * <p>
-     * Creates new QuizExerciseCache if required.
-     *
-     * @param quizExerciseId the id of the quiz exercise, must not be null
-     * @param writeOperation gets non-null and has  to return non-null.
-     */
-    private void performCacheWrite(Long quizExerciseId, UnaryOperator<QuizExerciseCache> writeOperation) {
-        log.info("Write quiz cache {}", quizExerciseId);
-        cachedQuizExercises.lock(quizExerciseId);
-        try {
-            cachedQuizExercises.set(quizExerciseId, writeOperation.apply(getTransientWriteCacheFor(quizExerciseId)));
-        }
-        finally {
-            cachedQuizExercises.unlock(quizExerciseId);
-        }
-    }
-
-    /**
-     * To perform the given action on QuizExerciseCache non-transient fields.
-     * <p>
-     * Will not execute the <code>writeOperation</code> if no QuizExerciseCache exists for the given id.
-     *
-     * @param quizExerciseId the id of the quiz exercise, must not be null
-     * @param writeOperation gets non-null and has  to return non-null.
-     */
-    private void performCacheWriteIfPresent(Long quizExerciseId, UnaryOperator<QuizExerciseCache> writeOperation) {
-        cachedQuizExercises.lock(quizExerciseId);
-        try {
-            QuizExerciseCache cachedQuiz = cachedQuizExercises.get(quizExerciseId);
-            if (cachedQuiz != null) {
-                log.info("Write quiz cache {}", quizExerciseId);
-                cachedQuizExercises.set(quizExerciseId, writeOperation.apply(cachedQuiz));
-            }
-        }
-        finally {
-            cachedQuizExercises.unlock(quizExerciseId);
-        }
-    }
-
-    /**
      * add a quizSubmission to the submissionHashMap
      *
      * @param quizExerciseId the quizExerciseId of the quiz the submission belongs to (first Key)
@@ -220,7 +116,7 @@ public class QuizScheduleService {
      */
     public void updateSubmission(Long quizExerciseId, String username, QuizSubmission quizSubmission) {
         if (quizSubmission != null && quizExerciseId != null && username != null) {
-            getTransientWriteCacheFor(quizExerciseId).getSubmissions().put(username, quizSubmission);
+            quizCache.getTransientWriteCacheFor(quizExerciseId).getSubmissions().put(username, quizSubmission);
         }
     }
 
@@ -234,7 +130,7 @@ public class QuizScheduleService {
     public void addResultForStatisticUpdate(Long quizExerciseId, Result result) {
         log.debug("add result for statistic update for quiz " + quizExerciseId + ": " + result);
         if (quizExerciseId != null && result != null) {
-            getTransientWriteCacheFor(quizExerciseId).getResults().put(result.getId(), result);
+            quizCache.getTransientWriteCacheFor(quizExerciseId).getResults().put(result.getId(), result);
         }
     }
 
@@ -246,7 +142,7 @@ public class QuizScheduleService {
      */
     private void addParticipation(Long quizExerciseId, StudentParticipation participation) {
         if (quizExerciseId != null && participation != null) {
-            getTransientWriteCacheFor(quizExerciseId).getParticipations().put(participation.getParticipantIdentifier(), participation);
+            quizCache.getTransientWriteCacheFor(quizExerciseId).getParticipations().put(participation.getParticipantIdentifier(), participation);
         }
     }
 
@@ -263,7 +159,7 @@ public class QuizScheduleService {
         if (quizExerciseId == null || username == null) {
             return null;
         }
-        QuizSubmission quizSubmission = getReadCacheFor(quizExerciseId).getSubmissions().get(username);
+        QuizSubmission quizSubmission = quizCache.getReadCacheFor(quizExerciseId).getSubmissions().get(username);
         if (quizSubmission != null) {
             return quizSubmission;
         }
@@ -282,7 +178,7 @@ public class QuizScheduleService {
         if (quizExerciseId == null || username == null) {
             return null;
         }
-        return getReadCacheFor(quizExerciseId).getParticipations().get(username);
+        return quizCache.getReadCacheFor(quizExerciseId).getParticipations().get(username);
     }
 
     /**
@@ -295,7 +191,7 @@ public class QuizScheduleService {
         if (quizExerciseId == null) {
             return null;
         }
-        QuizExercise quizExercise = getReadCacheFor(quizExerciseId).getExercise();
+        QuizExercise quizExercise = quizCache.getReadCacheFor(quizExerciseId).getExercise();
         if (quizExercise == null) {
             quizExercise = quizExerciseService.findOneWithQuestionsAndStatistics(quizExerciseId);
             if (quizExercise != null) {
@@ -311,13 +207,7 @@ public class QuizScheduleService {
      * @param quizExercise should include questions and statistics without Hibernate proxies!
      */
     public void updateQuizExercise(QuizExercise quizExercise) {
-        Objects.requireNonNull(quizExercise, "quizExercise must not be null");
-        cachedQuizExerciseUpdates.publish(quizExercise);
-    }
-
-    private void updateQuizExerciseLocally(QuizExercise quizExercise) {
-        log.debug("Quiz exercise {} updated in quiz exercise map: {}", quizExercise.getId(), quizExercise);
-        getTransientWriteCacheFor(quizExercise.getId()).setExercise(quizExercise);
+        quizCache.updateQuizExercise(quizExercise);
     }
 
     /**
@@ -371,7 +261,7 @@ public class QuizScheduleService {
                 log.info("Stop Quiz Schedule Service already disposed/cancelled");
                 // has already been disposed (sadly there is no method to check that)
             }
-            for (QuizExerciseCache cachedQuiz : cachedQuizExercises.values()) {
+            for (QuizExerciseCache cachedQuiz : quizCache.getAllQuizExerciseCaches()) {
                 if (cachedQuiz.getQuizStart() != null)
                     cancelScheduledQuizStart(cachedQuiz.getExerciseId());
             }
@@ -401,7 +291,7 @@ public class QuizScheduleService {
                 long delay = Duration.between(ZonedDateTime.now(), quizExercise.getReleaseDate()).toMillis();
                 var scheduledFuture = threadPoolTaskScheduler.schedule(new QuizStartTask(quizExerciseId), delay, TimeUnit.MILLISECONDS);
                 // save scheduled future in HashMap
-                performCacheWrite(quizExercise.getId(), quizExerciseCache -> {
+                quizCache.performCacheWrite(quizExercise.getId(), quizExerciseCache -> {
                     quizExerciseCache.setQuizStart(List.of(scheduledFuture.getHandler()));
                     return quizExerciseCache;
                 });
@@ -419,7 +309,7 @@ public class QuizScheduleService {
      * @param quizExerciseId the quiz exercise for which the quiz start should be canceled
      */
     public void cancelScheduledQuizStart(Long quizExerciseId) {
-        getReadCacheFor(quizExerciseId).getQuizStart().forEach(taskHandler -> {
+        quizCache.getReadCacheFor(quizExerciseId).getQuizStart().forEach(taskHandler -> {
             IScheduledFuture<?> scheduledFuture = threadPoolTaskScheduler.getScheduledFuture(taskHandler);
             try {
                 // if the task has been disposed, this will throw a StaleTaskException
@@ -438,7 +328,7 @@ public class QuizScheduleService {
                 // has already been disposed (sadly there is no method to check that)
             }
         });
-        performCacheWriteIfPresent(quizExerciseId, cachedQuiz -> {
+        quizCache.performCacheWriteIfPresent(quizExerciseId, cachedQuiz -> {
             cachedQuiz.setQuizStart(QuizExerciseCache.getEmptyQuizStartList());
             return cachedQuiz;
         });
@@ -448,7 +338,7 @@ public class QuizScheduleService {
      * Internal method to start and send the {@link QuizExercise} to the clients when called
      */
     void executeQuizStartNowTask(Long quizExerciseId) {
-        performCacheWriteIfPresent(quizExerciseId, quizExerciseCache -> {
+        quizCache.performCacheWriteIfPresent(quizExerciseId, quizExerciseCache -> {
             quizExerciseCache.getQuizStart().clear();
             log.debug("Removed quiz {} start tasks", quizExerciseId);
             return quizExerciseCache;
@@ -465,8 +355,7 @@ public class QuizScheduleService {
      * This will cause cached submissions, participations and results to be lost!
      */
     public void clearAllQuizData() {
-        cachedQuizExercises.values().forEach(QuizExerciseCache::clear);
-        cachedQuizExercises.clear();
+        quizCache.clear();
     }
 
     /**
@@ -476,10 +365,7 @@ public class QuizScheduleService {
      * @param quizExerciseId refers to one specific quiz exercise for which the data should be cleared
      */
     public void clearQuizData(Long quizExerciseId) {
-        QuizExerciseCache quizCache = cachedQuizExercises.remove(quizExerciseId);
-        if (quizCache != null) {
-            quizCache.clear();
-        }
+        quizCache.removeAndClear(quizExerciseId);
     }
 
     /**
@@ -499,9 +385,9 @@ public class QuizScheduleService {
         log.debug("Process cached quiz submissions");
         // global try-catch for error logging
         try {
-            for (Long quizExerciseId : cachedQuizExercises.keySet()) {
+            for (Long quizExerciseId : quizCache.getAllCachedQuizExerciseIds()) {
                 // this way near cache is used (values will deserialize new objects)
-                QuizExerciseCache cachedQuiz = cachedQuizExercises.get(quizExerciseId);
+                QuizExerciseCache cachedQuiz = quizCache.getCacheFor(quizExerciseId);
                 // safety check because of concurrent and distributed processes
                 if (cachedQuiz == null)
                     continue;
@@ -510,8 +396,7 @@ public class QuizScheduleService {
                 // check if quiz has been deleted
                 if (quizExercise == null) {
                     log.debug("Remove quiz " + quizExerciseId + " from resultHashMap");
-                    cachedQuizExercises.remove(quizExerciseId);
-                    cachedQuiz.clear();
+                    quizCache.removeAndClear(quizExerciseId);
                     continue;
                 }
 
@@ -606,7 +491,7 @@ public class QuizScheduleService {
 
     private void removeCachedQuiz(QuizExerciseCache cachedQuiz) {
         cancelScheduledQuizStart(cachedQuiz.getExerciseId());
-        cachedQuizExercises.remove(cachedQuiz.getExerciseId(), cachedQuiz);
+        quizCache.remove(cachedQuiz.getExerciseId());
     }
 
     private void sendQuizResultToUser(long quizExerciseId, StudentParticipation participation) {
